requirements.txt
asn1crypto==1.5.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.0.0
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
filelock==3.8.0
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
lightgbm==3.3.3
matplotlib==3.6.2
mlxtend==0.21.0
numpy==1.23.4
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pysnowflake==0.1.3
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
snowflake-connector-python==2.8.1
snowflake-snowpark-python==1.0.0
snowflake-sqlalchemy==1.4.3
SQLAlchemy==1.4.44
threadpoolctl==3.1.0
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.12
